# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "StudentPlanner.UI",
    "version": "1.0"
  },
  "paths": {
...
    "/event-requests": {
      "get": {
        "tags": [
          "EventRequests"
        ],
        "summary": "Retrieves event requests based on the role of the authenticated user.",
        "description": "Role-based behavior:\r\n- Admin: retrieves all event requests in the system.\r\n- Manager: retrieves only event requests created by the authenticated manager.",
        "responses": {
          "200": {
            "description": "Returns the list of event requests.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventRequestResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "EventRequests"
        ],
        "summary": "Creates a new event request.",
        "description": "Only managers are allowed to create event requests.\r\nThe newly created request is initially assigned a default status.",
        "requestBody": {
          "description": "The event request creation payload.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventRequestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "201": {
            "description": "Returns the newly created event request."
          }
        }
      }
    },
    "/event-requests/{eventId}": {
      "get": {
        "tags": [
          "EventRequests"
        ],
        "summary": "Retrieves a specific event request by its unique identifier.",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "description": "The unique identifier of the event request.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the requested event request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventRequestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "If the event request does not exist or is not accessible."
          }
        }
      }
    },
    "/event-requests/faculty": {
      "get": {
        "tags": [
          "EventRequests"
        ],
        "summary": "Retrieves event requests associated with a specific faculty.",
        "description": "Intended primarily for administrative use to filter event requests\r\nbelonging to a particular faculty.",
        "parameters": [
          {
            "name": "facultyId",
            "in": "query",
            "description": "The unique identifier of the faculty.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the filtered list of event requests.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventRequestResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "If faculty doesn't exist."
          }
        }
      }
    },
    "/event-requests/{id}": {
      "put": {
        "tags": [
          "EventRequests"
        ],
        "summary": "Updates the status of an existing event request.",
        "description": "Only administrators are allowed to update event request statuses.\r\nTypical status transitions include approval or rejection.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the event request.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The updated event request data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the updated request status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestStatus"
                }
              }
            }
          },
          "400": {
            "description": "If the request ID does not match the payload."
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "If the event request does not exist."
          }
        }
      },
      "delete": {
        "tags": [
          "EventRequests"
        ],
        "summary": "Deletes an existing event request.",
        "description": "Managers may only delete event requests they created.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the event request.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "204": {
            "description": "Event request deleted successfully."
          },
          "404": {
            "description": "If the event request does not exist."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateEventRequest": {
        "type": "object",
        "properties": {
          "details": {
            "$ref": "#/components/schemas/EventDetailsInputDTO"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "facultyId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EventDetailsInputDTO": {
        "required": [
          "endTime",
          "eventId",
          "location",
          "startTime",
          "title"
        ],
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "minLength": 1,
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EventDetailsOutputDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EventRequestResponse": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "requestType": {
            "$ref": "#/components/schemas/RequestType"
          },
          "requestStatus": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "submissionDate": {
            "type": "string",
            "format": "date-time"
          },
          "targetEventId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "details": {
            "$ref": "#/components/schemas/EventDetailsOutputDTO"
          },
          "managerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RequestStatus": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "RequestType": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "UpdateEventRequest": {
        "required": [
          "requestId",
          "requestStatus"
        ],
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "requestStatus": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}